requirements.txt
spoon-ai-sdk
spoon-toolkits
env
google-generativeai
spoon-toolkits
speechrecognition
elevenlabs
pyaudio
